♻️ refactor : fetchAPI 폴더 변경
