+++ src/app/items/[id]/page.tsx
@@ -1,7 +1,7 @@
 "use client";
-import fetchData from "@/app/api/fetchData";
 import RIOT_CONSTANT from "@/constants/RIOT_CONSTANT";
 import { Item } from "@/types/Items";
+import serverApi from "@/utils/serverApi";
 import parse from "html-react-parser";
 import Image from "next/image";
 
@@ -15,7 +15,7 @@ type ItemDetailType = Omit<Item, "colloq" | "into" | "tags" | "maps" | "stats">;
 
 const page = async ({ params }: Props) => {
   const id = params.id;
-  const data = await fetchData.fetchItems();
+  const data = await serverApi.fetchItems();
 
   const itemArray = data.filter((item) => item[0] === id)[0];
   const itemInfo: ItemDetailType = itemArray[1];
