+++ src/utils/serverApi.ts
@@ -1,6 +1,6 @@
+"use server";
 import RIOT_CONSTANT from "@/constants/RIOT_CONSTANT";
 import { ChampionDetail, GameData } from "@/types/ChampionDetail";
-import { RotationChampionData } from "@/types/Championrotation";
 import { Champion } from "@/types/Champions";
 import { Item, ItemData } from "@/types/Items";
 
@@ -31,25 +31,10 @@ const fetchItems = async (): Promise<[string, Item][]> => {
   return Object.entries(data);
 };
 
-//CSR
-const fetchRotationChampion = async (): Promise<
-  [number[], number[], number]
-> => {
-  const res = await fetch(`${RIOT_CONSTANT.BASE_URL}/api/rotationChampions`);
-  // const { data }: { data: RotationChampionData } = await res.json();
-  const data: RotationChampionData = await res.json();
-  return [
-    data.freeChampionIds,
-    data.freeChampionIdsForNewPlayers,
-    data.maxNewPlayerLevel,
-  ];
-};
-
-const fetchData = {
+const serverApi = {
   fetchChampion,
   fetchChampionDetail,
   fetchItems,
-  fetchRotationChampion,
 };
 
-export default fetchData;
+export default serverApi;
